Create MasterBootRecord
once the power button is turned on and the circuitry on the system gets power, the BIOS(basic input output system) is copied from ROM CHIP into the RAM . It initializes the hardware and performs the POST(power-on-self-test) , and performs legacy booting(which this project follows). the BIOS copies the first 512 bytes (MBR) of each bootable device ( flash memory in this project) into the memory location 0x7C00 to check for the 0xAA55 signature. It finds that signature and starts executing the code .